# Optimization: use write index cache in pop