Running benchmark on my machine:
without patch
```
SPSCQueue:
27176 ops/ms
193 ns RTT
boost::lockfree::spsc:
18873 ops/ms
298 ns RTT
```
with patch:
```
SPSCQueue:
41091 ops/ms
189 ns RTT
boost::lockfree::spsc:
18207 ops/ms
290 ns RTT
```