@bvbfan 2025-11-16
> @bvbfan did you compare your implementation with https://github.com/drogalis/SPSC-Queue ?
> 
> I'm curious how fast it is against yours.

```
dro::SPSCQueue: 
Mean: 23289 ops/ms 
Median: 23037 ops/ms 
Mean: 227 ns RTT 
Median: 226 ns RTT 

rigtorp::SPSCQueue:
Mean: 40904 ops/ms 
Median: 41492 ops/ms 
Mean: 197 ns RTT 
Median: 198 ns RTT
```